[Framework] Step, Sum and Gain blocks improved, bug in SiggnalBufferWriter fixed

+++ src/control/Sum.cpp
@@ -1,6 +1,6 @@
 #include <eeros/control/Sum.hpp>
 
-Sum::Sum(uint32_t nofInputs, sigdim_t dim) : out(dim), in(nofInputs), negated(nofInputs) {
+Sum::Sum(uint8_t nofInputs, sigdim_t dim) : out(dim), in(nofInputs), negated(nofInputs) {
 	
 }
 
@@ -8,7 +8,7 @@ void Sum::run() {
 	double sum;
 	for(sigdim_t i = 0; i < out.getDimension(); i++) {
 		sum = 0;
-		for(uint32_t j = 0; j < in.size(); j++) {
+		for(uint8_t j = 0; j < in.size(); j++) {
 			if(negated[j]) sum -= in[j].getValue(i);
 			else sum += in[j].getValue(i);
 		}
@@ -17,7 +17,7 @@ void Sum::run() {
 		out.setTimeStamp(in[0].getTimestamp(i));
 	}
 }
-RealSignalInput& Sum::getIn(uint32_t input) {
+RealSignalInput& Sum::getIn(uint8_t input) {
 	return in[input];
 }
 
@@ -25,6 +25,6 @@ RealSignalOutput& Sum::getOut() {
 	return out;
 }
 
-void Sum::negateInput(uint32_t input) {
+void Sum::negateInput(uint8_t input) {
 	negated[input] = true;
 }
